Refactoring MyCLabsEnum type
